Moved DashboardQueueAndTaskDetails into correct namespace. Improved namespace description
